Eliminate unnecessary updates of Variable Explorer
During work on a custom plugin for Spyder it has been noticed that there
are unnecessary updates of the Variable Explorer plugin when code is
debugged: Depending on circumstances the Variable Explorer is updated up
to 3 times. This patch eliminates the unnecessary updates.

In order to understand this patch, first some background information:
Whenever the debugger stops code execution, a command line interpreter
is started (cmd.Cmd.cmdloop()). The code looks roughly like:

    preloop()
    stop = False
    while not stop:
        line = input() # get user input (remote call into Spyder GUI)
        line = precmd(line)
        stop = onecmd(line) # interprets the command
        stop = postcmd(line)

Before this commit, both preloop() and postcmd() sent data to Spyder [0]
in order to update the Variable Explorer. The Variable Explorer also
requests an update of its data from the kernel when Spyder receives a
request for input for the debugger [1].

  (1) Assume that input() unblocks and the debugger receives a command
      to continue code execution (onecmd() returns True).
  (2) postcmd() from SpyderPdb in spyderpdb.py triggers update of
      Variable Explorer.
  (3) The command line interpreter exits and code execution is resumed.
  (4) When code execution is stopped, the command line interpreter is
      entered: preloop() triggers update of Variable Explorer.
  (5) Debugger requests input which in turn makes the Spyder GUI
      request an update for the Variable Explorer.
  === Timeline restarts at (1)
  (1) input() unblocks and the debugger receives a command to continue
      code execution (onecmd() returns True).
  (2) postcmd() from SpyderPdb in spyderpdb.py triggers update of
      Variable Explorer.
  (3) …

The update by preloop() (4) is not required due to the following update
(5) and the fact that the data can not (at least should not) change
between (4) and (5) because code execution is stopped.

The update by postcmd() (2) is not required because code execution is
about to be resumed (3).

The update by postcmd() (2) is also not required if the debugger, at
point (1), receives a command other than “continue code execution” [3].
In this case the command line input loop is not exited (onecmd() returns
False). The next request for input takes care of updating the Variable
Explorer as soon as the command has been processed [1].

This patch eliminates the unnecessary updates of the Variable Explorer
described above. This reduces the number of updates from 3 to 1 when the
debugger executes command which continue code execution, and from 2 to 1
for other commands.

The non-debugger case (“plain” IPython console) is not affected [1].

[0] By calling `SpyderPdb.notify_spyder()` which in turn calls
    `SpyderKernel.publish_pdb_state()`.

[1] The kernel’s request for input for the debugger is handled in Spyder
    by `DebuggingWidget.pdb_input()` [2]. Therein, a `executed` signal
    is emitted indicating that the previously entered command has been
    processed by the debugger. The Variable Explorer request a update of
    its data upon this signal.

    Updating of the Variable Explorer without a running debugger
    (“plain” IPython console) also relies on the `executed` signal. This
    signal is actually defined by QtConsole [4] and is emitted by the
    console when the kernel (`ipykernel`) finishes execution of code.

[2] plugins/ipythonconsole/widgets/debugging in Spyder’s source tree

[3] This applies e.g. to Python expressions: The debugger evaluates them
    and displays the resulting value.

[4] https://github.com/jupyter/qtconsole/blob/5.3.0/qtconsole/frontend_widget.py#L145-L147